# docs/swagger.yaml
openapi: 3.0.0
info:
  title: traQ API
  description: traQ v2 API
  version: 2.6.3
  license:
    name: MIT
    url: 'https://github.com/traPtitech/traQ/blob/master/LICENSE'
  contact:
    name: traP
    url: 'https://github.com/traPtitech/traQ'
servers:
  - url: https://q.trap.jp/api/1.0
    description: production environment
  - url: https://traq-dev.tokyotech.org/api/1.0
    description: staging environment
paths:
  /login:
    post:
      tags:
        - authentication
      operationId: login
      description: ログインを行います。リダイレクトパラメーターが存在する場合はログイン後にリダイレクトします
      parameters:
        - in: query
          name: redirect
          schema:
            type: string
          description: リダイレクト先
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
      responses:
        "204":
          description: 正常にログインできました。
        "302":
          description: 正常にログインできました。リダイレクトします。
        "400":
          description: ログインできませんでした。リクエスト内容が不正です
        "401":
          description: ログインできませんでした。認証情報が間違っています
        "403":
          description: ログインできませんでした。アカウントに問題があります
  /logout:
    post:
      tags:
        - authentication
      operationId: logout
      description: ログアウトを行います。リダイレクトパラメーターが存在する場合はログアウト後にリダイレクトします
      parameters:
        - in: query
          name: redirect
          schema:
            type: string
          description: リダイレクト先
      security: []
      responses:
        "204":
          description: 正常にログアウトできました。
        "302":
          description: 正常にログアウトできました。リダイレクトします。
  /public/icon/{username}:
    get:
      tags:
        - public
      operationId: getPublicUserIcon
      description: 指定したユーザーのアイコン画像を取得します。
      parameters:
        - name: username
          description: 画像を取得するユーザーのユーザー名
          in: path
          required: true
          schema:
            type: string
      security: []
      responses:
        "200":
          description: 正常に取得できました。
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/gif:
              schema:
                type: string
                format: binary
        "404":
          description: 取得できませんでした。指定したユーザーは存在しません。

  # /public/emoji.json

  # /public/emoji.css

  # /public/emoji/{stampID}
  /oauth2/authorize:
    get:
      tags:
        - authorization
      operationId: getOauth2Authorize
      description: OAuth2 認可エンドポイント
      parameters:
        - name: response_type
          in: query
          schema:
            $ref: "#/components/schemas/OAuth2ResponseType"
        - name: client_id
          in: query
          required: true
          schema:
            type: string
        - name: redirect_uri
          in: query
          schema:
            type: string
        - name: scope
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: code_challenge
          in: query
          schema:
            type: string
        - name: code_challenge_method
          in: query
          schema:
            type: string
        - name: nonce
          in: query
          schema:
            type: string
        - name: prompt
          in: query
          schema:
            $ref: "#/components/schemas/OAuth2Prompt"
      security: []
      responses:
        "302":
          description: 結果に応じてリダイレクトします。
        "400":
          description: リクエストが不正です。
        "403":
          description: リクエストが許可されていません。
    post:
      tags:
        - authorization
      operationId: postOauth2Authorize
      description: OAuth2 認可エンドポイント
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/OAuth2Authorization"
      responses:
        "302":
          description: 結果に応じてリダイレクトします。
        "400":
          description: リクエストが不正です。
        "403":
          description: リクエストが許可されていません。
  /oauth2/authorize/decide:
    post:
      tags:
        - authorization
      operationId: postOauth2AuthorizeDecide
      description: OAuth2 認可承諾
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/OAuth2Decide"
      responses:
        "302":
          description: RedirectURIに結果をリダイレクトします。
        "400":
          description: リクエストが不正です。
        "403":
          description: リクエストが許可されていません。
  /oauth2/token:
    post:
      tags:
        - authorization
      operationId: postOauth2Token
      description: OAuth2 トークンエンドポイント
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/PostOAuth2Token"
      responses:
        "200":
          description: トークンが正常に発行されました。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuth2Token"
        "400":
          description: トークン発行に失敗しました。
        "403":
          description: トークン発行に失敗しました。

  # /oauth2/revoke

  # /metrics
  /users:
    get:
      tags:
        - user
      operationId: getUsers
      description: 全ユーザーのリストを取得します。
      security:
        - traqOAuth2:
            - read
            - manage_bot
      responses:
        "200":
          description: |+
            正常に取得できました。
            ユーザーリストを返します。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
    post:
      tags:
        - user
      operationId: registerUser
      description: ユーザー登録します
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegister"
      responses:
        "201":
          description: 正常に登録できました。
        "400":
          description: 登録できませんでした。リクエスト内容が不正です。
        "403":
          description: 登録できませんでした。権限がありません。
  /users/me:
    get:
      tags:
        - user
      operationId: getMe
      description: 自分のユーザー情報を取得します。
      security:
        - traqOAuth2:
            - read
            - manage_bot
      responses:
        "200":
          description: 正常に取得できました。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
    patch:
      tags:
        - user
      operationId: changeMe
      description: 自分のユーザー情報を変更します。
      security:
        - traqOAuth2:
            - write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserData"
      responses:
        "204":
          description: 正常に変更できました。
        "400":
          description: 正常に変更できませんでした。リクエスト内容が不正です。
...
  /users/me/icon:
    get:
      tags:
        - user
      operationId: getMyIcon
      description: 自分のアイコン画像を取得します。
      security:
        - traqOAuth2:
            - read
      responses:
        "200":
          description: 正常に取得できました。
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/gif:
              schema:
                type: string
                format: binary
    put:
      tags:
        - user
      operationId: changeMyIcon
      description: 自分のアイコンを更新します。
      security:
        - traqOAuth2:
            - write
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              properties:
                file:
                  type: string
                  format: binary
                  description: 1MBまでのpng, jpeg, gif
            encoding:
              file:
                contentType: image/png, image/jpeg, image/gif
      responses:
        "200":
          description: 正常に更新できました。
        "400":
          description: 正常に更新できませんでした。不正なファイルです。
...
  /users/me/unread/channels:
    get:
      tags:
        - unread
      operationId: getUnreadChannels
      description: 未読チャンネル情報のリストを取得します。
      security:
        - traqOAuth2:
            - read
      responses:
        "200":
          description: 正常に取得できました。
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UnreadChannel"
...
  /users/{userID}:
    parameters:
      - $ref: "#/components/parameters/userIdInPath"
    get:
      tags:
        - user
      operationId: getUser
      description: ユーザーの詳細を取得します。
      security:
        - traqOAuth2:
            - read
            - manage_bot
      responses:
        "200":
          description: |+
            正常に取得できました。
            ユーザー情報を返します。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
        "404":
          description: +|
            正常に取得できませんでした。
            指定したユーザーは存在しません。

  # PATCH /users/{userID}

  # /users/{userID}/status

  # /users/{userID}/password
  /users/{userID}/messages:
    parameters:
      - $ref: "#/components/parameters/userIdInPath"
    get:
      tags:
        - message
      operationId: getDirectMessages
      description: DMチャンネルに存在するメッセージを取得します。
      security:
        - traqOAuth2:
            - read
      parameters:
        - $ref: "#/components/parameters/limitInQuery"
        - $ref: "#/components/parameters/offsetInQuery"
        - $ref: "#/components/parameters/sinceInQuery"
        - $ref: "#/components/parameters/untilInQuery"
        - $ref: "#/components/parameters/inclusiveInQuery"
        - $ref: "#/components/parameters/orderInQuery"
      responses:
        "200":
          description: |+
            正常に取得ができました。
            メッセージの配列を返します。
          headers:
            X-TRAQ-MORE:
              $ref: "#/components/headers/X-TRAQ-MORE"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageList"
        "400":
          description: |+
            変更に失敗しました
            不正なリクエストです
        "404":
          description: |+
            取得に失敗しました。
            指定したユーザーは存在しません。
    post:
      tags:
        - message
      operationId: postDirectMessage
      description: DMチャンネルにメッセージを投稿します。
      parameters:
        - name: embed
          description: メンション・チャンネルリンクを自動埋め込みする場合に1を指定する
          in: query
          required: false
          schema:
            type: integer
      security:
        - traqOAuth2:
            - write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessage"
      responses:
        "201":
          description: |+
            投稿に成功しました。
            投稿されたメッセージが返されます。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "404":
          description: +|
            投稿に失敗しました。
            指定したユーザーは存在しません。
...
  /users/{userID}/groups:
    parameters:
      - $ref: "#/components/parameters/userIdInPath"
    get:
      tags:
        - user group
      operationId: getUserGroups
      description: 所属するユーザーグループのIDを取得します
      security:
        - traqOAuth2:
            - read
      responses:
        "200":
          description: 正常に取得できました
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UUIDs"
        "404":
          description: 正常に取得できませんでした。指定したユーザーは存在しません。
...
  /notification:
    get:
      tags:
        - notification
      operationId: connectNotification
      description: 通知ストリーム(Server Sent Events)に接続します。
      security:
        - traqOAuth2:
            - read
      responses:
        "200":
          description: 接続が完了しました。
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: r_session
    traqOAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /api/1.0/oauth2/authorize
          tokenUrl: /api/1.0/oauth2/token
          scopes:
            read: 読み取りスコープ
            write: 書き込みスコープ
            manage_bot: bot関連読み書きスコープ
  parameters:
...
    userIdInPath:
      name: userID
      description: 操作の対象となるユーザーID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    OAuth2ResponseType:
      type: string
      enum:
        - code
        - token
        - none
    OAuth2Prompt:
      type: string
      enum:
        - none
...
    OAuth2Decide:
      type: object
      required:
        - submit
      properties:
        submit:
          type: string
          description: 承諾する場合は"approve"
    UserLogin:
      type: object
      required:
        - name
        - pass
      properties:
        name:
          type: string
        pass:
          type: string
          format: password
...
    UserList:
      type: array
      items:
        $ref: "#/components/schemas/User"
...
security:
  - cookieAuth: []